requirements.txt
Django==1.10.6
ipdb==0.10.2
ipython==2.4.1
psycopg2==2.6.1
Pillow
